Launcher: add notes about changes to unhandled exception filter
The engine executable has been patched to no longer install this, therefore this function will never be called as we install a vectored exception filter anyways. This solves the issue where the crash message application gets called more than once when a crash occurs before the SDK detours are initialized.